cloud-functions/login_user/requirements.txt
functions-framework==3.*
PyJWT>=2.10.1
pymongo>=4.15.1
bcrypt>=4.0.1
requests